비기능: '각 플레이어가 베팅 금액을 입력하는 기능'에 대한 에러처리.

+++ src/main/java/domain/user/PlayerFactory.java
@@ -2,7 +2,7 @@ package domain.user;
 
 
 public class PlayerFactory {
-    public Player create(String name, int bettingMoney) {
+    public Player create(String name, double bettingMoney) {
         return new Player(name, bettingMoney);
     }
 }
